requirements.txt
alembic==1.6.5
bcrypt==3.2.0
bleach==3.3.0
cffi==1.14.5
click==8.0.1
colorama==0.4.4
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==20.9
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.20
webencodings==0.5.1
Werkzeug==2.0.1
